requirement.txt
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
beautifulsoup4==4.11.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==36.0.2
decorator==4.4.2
ffmpeg-python==0.2.0
future==0.18.2
h11==0.13.0
idna==3.3
imageio==2.18.0
imageio-ffmpeg==0.4.7
lxml==4.8.0
moviepy==1.0.3
msedge-selenium-tools==3.141.4
numpy==1.22.3
outcome==1.1.0
Pillow==9.1.0
proglog==0.1.9
pycodestyle==2.8.0
pycparser==2.21
pyOpenSSL==22.0.0
pyperclip==1.8.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
PyQt6==6.3.0
PyQt6-Qt6==6.3.0
PyQt6-sip==13.3.1
PySocks==1.7.1
pytube==12.0.0
requests==2.27.1
selenium==3.141.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
toml==0.10.2
tqdm==4.64.0
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
wsproto==1.1.0
